IoT_API/requirements.txt
alembic==1.7.6
atomicwrites==1.4.0
attrs==21.4.0
click==8.0.4
colorama==0.4.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pymysql
pytest==7.1.0
SQLAlchemy
tomli==2.0.1
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
cryptography
requests
thrift>=0.13
apache-iotdb